add dash example of slow callback
